MacGui: fixed some audio presets key: AudioSamplerate -> string, AudioTrackDRCSlider -> float.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6282 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBAudioTrackPreset.h
@@ -24,7 +24,7 @@
 @property (nonatomic, readwrite) int bitRate;
 
 @property (nonatomic, readwrite) int gain;
-@property (nonatomic, readwrite) int drc;
+@property (nonatomic, readwrite) float drc;
 
 /**
  *  Arrays of possible options for the track properties.
